payment tested

+++ src/main/java/com/miu/onlinemarketplace/utils/Utility.java
@@ -1,7 +1,7 @@
 package com.miu.onlinemarketplace.utils;
 
 public class Utility {
-    public static final int TAX = 7;
+    public static final double TAX = 7;
 
     public static final int SHIPPING_CHARGE = 2;
